+++ packages/nodes-base/nodes/Contentful/GenericFunctions.ts
